scraper/requirements.txt
playwright
beautifulsoup4
langchain-google-genai
chromadb
httpx
python-dotenv
